pyproject.toml
[project]
name = "inference-autopilot"
version = "0.1.0"
description = "Self-optimizing agentic inference system for Nebius Token Factory"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "websockets>=13.0",
    "numpy>=1.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
